bdd/requirements.txt
pytest
pytest-bdd
playwright
pytest-html
pytest-playwright
requests
python-dotenv